Change install/uninstall to be case insensitive
Searching modules is already case insensitive, however when trying to
install or uninstall modules it requires the case entered to be exactly as
it shows in the module list.

Per the CKAN spec ( https://github.com/KSP-CKAN/CKAN/blob/master/Spec.md ),
identifiers "must be both: case sensitive for machines, and unique regardless
of capitalization for human consumption and case-ignorant systems."

Switching to case insensitive matching for install/uninstall will improve
usability, and the spec ensures that doing so will not cause issues
with name collisions.